Create a normal buffer if none are listed
vim-bbye provides a `:Bdelete` command that closes the buffer but not
the window with it. When last buffer is closed it creates a scratchpad
buffer that has things like `nobuflisted` set. This throws off
buffingwheel because it sees 0 buffers listed. Now if no buffers are
listed it will create a normal buffer.